style(css): 잘못된 속성 수정
