+++ components/README.md
@@ -1,35 +1,35 @@
-# Button Module
+## Modal 컴포넌트
 
 ### 기능 구현 목록
 
-- [ ] 모달 컴포넌트 제작
-- [ ] 모바일에서 사용 가능한 모달 컴포넌트
+- [x] 모달 컴포넌트 제작
+- [x] 모바일에서 사용 가능한 모달 컴포넌트
 
 - props
 
-  - [ ] position : 모달 위치 ( 중앙, 하단 )
-  - [ ] title : 모달의 제목
-  - [ ] close 버튼 : icon, button, none
-  - [ ] confirm 버튼 : text string 만 전달받음
-  - [ ] 사용자 정의 이벤트 핸들러를 지원
-    - [ ] 모달 열기, 닫기, 확인
+  - [x] position : 모달 위치 ( 중앙, 하단 )
+  - [x] title : 모달의 제목
+  - [x] close 버튼 : icon, button, none
+  - [x] confirm 버튼 : text string 만 전달받음
+  - [x] 사용자 정의 이벤트 핸들러를 지원
+  - [x] 모달 열기, 닫기, 확인
 
 - 배포
-  - [ ] 모달 컴포넌트를 npm으로 배포
+  - [x] 모달 컴포넌트를 npm으로 배포
 
 ### 기능 구현 목록
 
-- [ ] 모달 컴포넌트 제작
-- [ ] 모바일에서 사용 가능한 모달 컴포넌트
+- [x] 모달 컴포넌트 제작
+- [x] 모바일에서 사용 가능한 모달 컴포넌트
 
 - props
 
-  - [ ] position : 모달 위치 ( 중앙, 하단 )
-  - [ ] title : 모달의 제목
-  - [ ] close 버튼 : icon, button, none
-  - [ ] confirm 버튼 : text string 만 전달받음
-  - [ ] 사용자 정의 이벤트 핸들러를 지원
-    - [ ] 모달 열기, 닫기, 확인
+  - [x] position : 모달 위치 ( 중앙, 하단 )
+  - [x] title : 모달의 제목
+  - [x] close 버튼 : icon, button, none
+  - [x] confirm 버튼 : text string 만 전달받음
+  - [x] 사용자 정의 이벤트 핸들러를 지원
+    - [x] 모달 열기, 닫기, 확인
 
 - 배포
-  - [ ] 모달 컴포넌트를 npm으로 배포
+  - [x] 모달 컴포넌트를 npm으로 배포
