+++ components/src/lib/Input/Input.style.ts
@@ -15,3 +15,11 @@ export const InputBox = styled.input<{ $isError: boolean }>`
     border: 1px solid black;
   }
 `;
+
+export const FlexBox = styled.div`
+  display: flex;
+  gap: 5px;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+`;
