+++ components/src/lib/Modal/AlertModal/AlertModal.tsx
@@ -1,24 +1,24 @@
-import Modal, { ModalPosition, ModalSize } from "../Modal/Modal";
+import Modal, { ModalPosition, ModalSize } from "../Modal";
 
 export interface AlertModalProps {
   isOpen: boolean;
   onClose: () => void;
   onConfirm: () => void;
   title: string;
-  message: string;
   buttonText?: string;
   position?: ModalPosition;
   modalSize?: ModalSize;
+  children?: React.ReactNode;
 }
 const AlertModal = ({
   isOpen,
   onClose,
   onConfirm,
   title,
-  message,
   modalSize = "medium",
   buttonText = "확인",
   position = "center",
+  children,
 }: AlertModalProps) => {
   return (
     <Modal
@@ -28,7 +28,7 @@ const AlertModal = ({
       size={modalSize}
     >
       <Modal.Title>{title}</Modal.Title>
-      <Modal.Content>{message}</Modal.Content>
+      <Modal.Content>{children}</Modal.Content>
       <Modal.Footer>
         <Modal.StyledButton
           onClickEvent={onConfirm}
