+++ components/src/lib/Modal/Modal.style.ts
@@ -1,8 +1,35 @@
 import styled, { css } from "styled-components";
+import { ModalContentPosition } from "./Modal";
 
-export type ModalPosition = "center" | "bottom";
+export type ModalPosition = "center" | "bottom" | "top";
 
 export type ModalSize = "small" | "medium" | "large" | "full";
+
+export const ModalBottomStyle = css`
+  @media (max-width: 567px) {
+    width: 100%;
+    top: auto;
+    transform: translate(-50%);
+    bottom: 0;
+    border-radius: 8px 8px 0px 0px;
+  }
+`;
+
+export const ModalCenterStyle = css`
+  @media (max-width: 567px) {
+    transform: translate(-50%, -50%);
+    top: 50%;
+  }
+`;
+
+export const ModalTopStyle = css`
+  @media (max-width: 567px) {
+    width: 100%;
+    top: 0;
+    transform: translate(-50%);
+    border-radius: 0px 0px 8px 8px;
+  }
+`;
 
 export const MODAL_WIDTH_MAP: Record<ModalSize, string> = {
   small: "320px",
@@ -42,22 +69,13 @@ export const ModalOuter = styled.div<{
   min-height: 150px;
   width: ${({ $size: size }) => MODAL_WIDTH_MAP[size]};
 
+  top: 50%;
+  transform: translate(-50%, -50%);
+  border-radius: 8px;
   ${({ $position }) => {
-    if ($position === "bottom") {
-      return css`
-        top: auto;
-        transform: translate(-50%);
-        bottom: 0;
-        border-radius: 8px 8px 0px 0px;
-      `;
-    }
-    if ($position === "center") {
-      return css`
-        top: 50%;
-        transform: translate(-50%, -50%);
-        border-radius: 8px;
-      `;
-    }
+    if ($position === "bottom") return ModalBottomStyle;
+    if ($position === "center") return ModalCenterStyle;
+    return ModalTopStyle;
   }};
 `;
 
@@ -66,7 +84,7 @@ export const ModalInner = styled.div`
   padding: 30px;
   display: flex;
   flex-direction: column;
-  justify-content: flex-start;
+  justify-content: "flex-left";
 `;
 
 export const Title = styled.span`
@@ -97,14 +115,17 @@ export const CloseIcon = styled.button`
   color: black;
 `;
 
-export const Content = styled.div`
+export const Content = styled.div<{ contentPosition: ModalContentPosition }>`
   margin-bottom: 10px;
   text-align: left;
   display: flex;
   align-items: center;
-  justify-content: left;
+  justify-content: ${({ contentPosition }) =>
+    contentPosition === "center" ? "center" : "flex-start"};
 `;
 
-export const ButtonWrapper = styled.div`
+export const Footer = styled.div`
   display: flex;
+  justify-content: end;
+  gap: 10px;
 `;
