--- components/src/lib/Modal/Modal/Modal.style.ts
@@ -1,131 +1,0 @@
-import styled, { css } from "styled-components";
-import { ModalContentPosition } from "./Modal";
-
-export type ModalPosition = "center" | "bottom" | "top";
-
-export type ModalSize = "small" | "medium" | "large" | "full";
-
-export const ModalBottomStyle = css`
-  @media (max-width: 567px) {
-    width: 100%;
-    top: auto;
-    transform: translate(-50%);
-    bottom: 0;
-    border-radius: 8px 8px 0px 0px;
-  }
-`;
-
-export const ModalCenterStyle = css`
-  @media (max-width: 567px) {
-    transform: translate(-50%, -50%);
-    top: 50%;
-  }
-`;
-
-export const ModalTopStyle = css`
-  @media (max-width: 567px) {
-    width: 100%;
-    top: 0;
-    transform: translate(-50%);
-    border-radius: 0px 0px 8px 8px;
-  }
-`;
-
-export const MODAL_WIDTH_MAP: Record<ModalSize, string> = {
-  small: "320px",
-  medium: "480px",
-  large: "600px",
-  full: "100vw",
-};
-
-export const ModalWrapper = styled.div<{ open: boolean }>`
-  position: fixed;
-  display: ${({ open }) => (open ? "flex" : "none")};
-  z-index: 10;
-`;
-
-export const ModalBackground = styled.div`
-  position: fixed;
-  width: 100vw;
-  height: 100vh;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  background: rgba(0, 0, 0, 0.35);
-`;
-
-export const ModalOuter = styled.div<{
-  $position: ModalPosition;
-  $size: ModalSize;
-}>`
-  position: fixed;
-  left: 50%;
-  background-color: white;
-  color: black;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  min-height: 150px;
-  width: ${({ $size: size }) => MODAL_WIDTH_MAP[size]};
-
-  top: 50%;
-  transform: translate(-50%, -50%);
-  border-radius: 8px;
-  ${({ $position }) => {
-    if ($position === "bottom") return ModalBottomStyle;
-    if ($position === "center") return ModalCenterStyle;
-    return ModalTopStyle;
-  }};
-`;
-
-export const ModalInner = styled.div`
-  width: 100%;
-  padding: 30px;
-  display: flex;
-  flex-direction: column;
-  justify-content: "flex-left";
-`;
-
-export const Title = styled.span`
-  font-size: 24px;
-  font-weight: bold;
-  margin-bottom: 16px;
-  text-align: left;
-`;
-
-export const ModalHeader = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-start;
-`;
-
-export const CloseIcon = styled.button`
-  width: 30px;
-  height: 30px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 0;
-  cursor: pointer;
-  position: absolute;
-  top: 20px;
-  right: 20px;
-  background-color: transparent;
-  color: black;
-`;
-
-export const Content = styled.div<{ contentPosition: ModalContentPosition }>`
-  margin-bottom: 10px;
-  text-align: left;
-  display: flex;
-  align-items: center;
-  justify-content: ${({ contentPosition }) =>
-    contentPosition === "center" ? "center" : "left"};
-`;
-
-export const Footer = styled.div`
-  display: flex;
-  justify-content: end;
-  gap: 10px;
-`;
