+++ components/src/lib/index.ts
@@ -1,4 +1,4 @@
-export { default as Modal } from "@/lib/Modal/Modal/Modal";
+export { default as Modal } from "@/lib/Modal/Modal";
 export { default as Button } from "@/lib/Button/Button";
 export { default as AlertModal } from "@/lib/Modal/AlertModal/AlertModal";
 export { default as ConfirmModal } from "@/lib/Modal/ConfirmModal/ConfirmModal";
